+++ priv/repo/migrations/20250407072552_add_note_change_fields_to_structure_notes.exs
@@ -1,0 +1,9 @@
+defmodule TuApp.Repo.Migrations.AddLastChangedFieldsToStructureNotes do
+  use Ecto.Migration
+
+  def change do
+    alter table(:structure_notes) do
+      add :last_changed_by, :bigint, null: true
+    end
+  end
+end