+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - [TD-7192] Use whitespace tokenizer with word delimiter filter to ignore `_` on text search
+- [TD-7141] Added and indexed last changed field for structure notes
 
 ## [Unreleased]